Increment ad ids by 1000000
Ad posters are presented in a landscape format - 256x360 px - instead of the normal 360x256 px in portrait.

Should we be on v770 and metadata is requested for an id greater than 1000000, we'll subtract such and return the same in the database. This allows us to maintain the same metadata yet have separate assets.